+++ userprovided/__init__.py
@@ -5,10 +5,11 @@ from userprovided import cloud
 from userprovided import date
 from userprovided import hash
 from userprovided import mail
+from userprovided import parameters
 from userprovided import port
 from userprovided import url
 
 
 name = "userprovided"
-__version__ = "0.5.4"
+__version__ = "0.6.0"
 __author__ = "Rüdiger Voigt"